requirements.txt
FreeSimpleGUI
ytmusicapi
python-vlc
yt-dlp
